+++ lib/moleculer/support.rb
@@ -1,4 +1,6 @@
-require_relative "support/hash_util"
+# frozen_string_literal: true
+
+require_relative "support/hash"
 require_relative "support/string_util"
 require_relative "support/open_struct"
 require_relative "support/log_proxy"